+++ test/RegionalLCIASpec.hs
@@ -24,6 +24,7 @@ import Method.Types (
     Location (..),
     MethodCF (..),
  )
+import TestHelpers (unitDef)
 import Types (
     Activity (..),
     BiosphereFlow (..),
@@ -36,7 +37,7 @@ import Types (
     emptyProductIndex,
  )
 import qualified Types as VT
-import UnitConversion (UnitConfig (..), UnitDef (..), defaultUnitConfig, mkUnitConfig)
+import UnitConversion (UnitConfig (..), defaultUnitConfig, mkUnitConfig)
 
 -- ---------------------------------------------------------------------------
 -- Fixture
@@ -62,7 +63,7 @@ kgUnitConfig :: UnitConfig
 kgUnitConfig =
     mkUnitConfig
         ["mass", "length", "time", "energy", "area", "volume", "count", "currency"]
-        (M.fromList [("kg", UnitDef [1, 0, 0, 0, 0, 0, 0, 0] 1.0)])
+        (M.fromList [("kg", unitDef "mass" 1.0)])
 
 testFlow :: BiosphereFlow
 testFlow =